+++ common/src/res_time.rs
@@ -1,5 +1,5 @@
 pub struct TimeRes {
-    pub tick: u32,
+    pub tick: u64,
 }
 impl Default for TimeRes {
     fn default() -> Self {
